Place the encrypted data key first in the EncryptedRecord
The intent of this change is to allow for "data key locality".
That is, if we sort a collection of serialised `EncryptedRecord`s, those which use the same data key will be adjacent.
The benefit of this is that when using a key store which has an expensive data key decryption operation, caching decrypted data keys will have a significant performance impact.